+++ next.config.js
@@ -1,8 +1,8 @@
 /** @type {import('next').NextConfig} */
 
-module.exports  = {
-  experimental: {appDir: true}, 
+module.exports = {
+  experimental: { appDir: true },
   images: {
-    domains: ['cdn.simpleicons.org', 'res.cloudinary.com']
-  }
+    domains: ['cdn.simpleicons.org', 'res.cloudinary.com', 'i.scdn.co'],
+  },
 };