+++ src/app/[locale]/callback/page.tsx
@@ -1,0 +1,14 @@
+import React, { FC } from 'react';
+
+interface ICallbackProps {}
+
+const Callback: FC<ICallbackProps> = ({}) => {
+  return (
+    <div>
+      <h1>Callback</h1>
+      <p>Callback page content</p>
+    </div>
+  );
+};
+
+export default Callback;