+++ src/utils/constants/times.ts
@@ -1,0 +1,16 @@
+export enum ETime {
+  SECOND = 1000,
+  HALF_SECOND = 500,
+  MINUTE = 60 * 1000,
+  HALF_MINUTE = 30 * 1000,
+  HOUR = 60 * 60 * 1000,
+  HALF_HOUR = 30 * 60 * 1000,
+  DAY = 24 * 60 * 60 * 1000,
+  HALF_DAY = 12 * 60 * 60 * 1000,
+  WEEK = 7 * 24 * 60 * 60 * 1000,
+  HALF_WEEK = 3 * 24 * 60 * 60 * 1000,
+  MONTH = 30 * 24 * 60 * 60 * 1000,
+  HALF_MONTH = 15 * 24 * 60 * 60 * 1000,
+  YEAR = 365 * 24 * 60 * 60 * 1000,
+  HALF_YEAR = 182 * 24 * 60 * 60 * 1000,
+}