+++ CHANGELOG.md
@@ -3,6 +3,11 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## 2.5.2 (2023-08-14)
+
+- Added Spotify widget to the project to show the current song that I'm listening to.
+- Addes Highligth.js to the project to show code snippets in a better way.
+
 ## 2.5.1 (2023-08-04)
 
 - Reduced size of flags thumbnails on the language selector